# fast setup for output tensor in tensor iterator